homeWork5 update type variables

+++ src/main/java/ru/otus/javabasic/hw5/animals/Cat.java
@@ -1,7 +1,7 @@
 package ru.otus.javabasic.hw5.animals;
 
 public class Cat extends Animal {
-    public Cat(String name, int runSpeed, int swimSpeed, int stamina) {
+    public Cat(String name, float runSpeed, float swimSpeed, float stamina) {
         super(name, runSpeed, swimSpeed, stamina);
         swimFactor = 0;
     }
