+++ src/main/java/ru/otus/javabasic/hw5/animals/Dog.java
@@ -1,11 +1,11 @@
 package ru.otus.javabasic.hw5.animals;
 
 public class Dog extends Animal {
-    public Dog(String name, int speedRun, int factorRun, int speedSwim, int factorSwim, int stamina) {
+    public Dog(String name, float speedRun, float factorRun, float speedSwim, float factorSwim, float stamina) {
         super(name, speedRun, factorRun, speedSwim, factorSwim, stamina);
     }
 
-    public Dog(String name, int runSpeed, int swimSpeed, int stamina) {
+    public Dog(String name, float runSpeed, float swimSpeed, float stamina) {
         super(name, runSpeed, swimSpeed, stamina);
         swimFactor = 2;
     }
